migrate serverinfo as info_server
