+++ discord-frontend/hartex-discord-commands/src/utilities/mod.rs
@@ -27,4 +27,3 @@
 //! - serverinfo
 
 pub mod info;
-pub mod serverinfo;
