+++ localization/locales/zh-TW/discord-frontend/commands/utilities-plugin.ftl
@@ -36,4 +36,3 @@ serverinfo-embed-memberinfo-field-name=成員資訊
 serverinfo-embed-memberinfo-membercount-subfield-name=成員數：
 serverinfo-embed-memberinfo-humancount-subfield-name=真人數：
 serverinfo-embed-memberinfo-botcount-subfield-name=機器人數：
-serverinfo-server-not-supported=目前不支援查詢特定伺服器的資訊。
